test: move hash_code integration tests to integration directory
Following PR feedback, I have moved the end-to-end integration tests for
`hash_code` to `tests/integration/test_hash_code_integration.py`.
Pure unit tests for `digest()` and `CallLookup` remain in `tests/unit`.
